pyproject.toml
[tool.poetry]
name = "personal-journal"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flake8 = "^7.1.1"
black = "^24.10.0"
isort = "^5.13.2"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
pillow = "^11.1.0"
djangorestframework-simplejwt = "^5.4.0"
django-filter = "^24.3"
coverage = "^7.6.10"
drf-yasg = "^1.21.8"
django = "4.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
